requirements.txt
flask
yt-dlp
openai-whisper 